study-agents-differences/requirements.txt
agno==1.1.4
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.3
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1; platform_python_implementation == "PyPy"
charset-normalizer==3.4.1
click==8.1.8
cohere==5.13.12
colorama==0.4.6
dataclasses-json==0.6.7
deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
docstring-parser==0.16
eval-type-backport==0.2.2
fastavro==1.10.0
filelock==3.17.0
filetype==1.2.0
frozenlist==1.5.0
fsspec==2025.2.0
gitdb==4.0.12
gitpython==3.1.44
google-auth==2.38.0
greenlet==3.1.1
griffe==1.5.7
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
idna==3.10
importlib-metadata==8.6.1
jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain-core==0.3.37
langchain-openai==0.3.6
langgraph==0.2.74
langgraph-checkpoint==2.0.16
langgraph-sdk==0.1.53
langsmith==0.3.8
litellm==1.61.11
llama-cloud==0.1.13
llama-cloud-services==0.6.1
llama-index==0.12.19
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.0
llama-index-core==0.12.19
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.8
llama-index-llms-openai==0.3.20
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.5
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.1
logfire-api==3.6.1
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mistralai==1.5.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.3
openai==1.63.2
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.2.1
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22; platform_python_implementation == "PyPy"
pydantic==2.10.6
pydantic-ai==0.0.24
pydantic-ai-slim[anthropic,cohere,groq,mistral,openai,vertexai]==0.0.24
pydantic-core==2.27.2
pydantic-graph==0.0.24
pydantic-settings==2.7.1
pygments==2.19.1
pypdf==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
pyyaml==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
sqlalchemy[asyncio]==2.0.38
striprtf==0.0.26
tavily-python==0.5.1
tenacity==9.0.0
tiktoken==0.9.0
tokenizers==0.21.0
tomli==2.2.1
tqdm==4.67.1
typer==0.15.1
types-requests==2.32.0.20241016
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2025.1
urllib3==2.3.0
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
